Move NullJsonField to corehq.util.models
